+++ crates/miden-tx/src/errors/mod.rs
@@ -67,8 +67,6 @@ pub enum TransactionExecutorError {
     // It is boxed to avoid triggering clippy::result_large_err for functions that return this type.
     #[error("transaction is unauthorized with summary {0:?}")]
     Unauthorized(Box<TransactionSummary>),
-    #[error("transaction returned unauthorized event but a commitment did not match: {0}")]
-    TransactionSummaryCommitmentMismatch(Box<str>),
 }
 
 // TRANSACTION PROVER ERROR